[#1074]: Align names of protocol adapter main application classes
