+++ adapters/http-vertx/src/main/java/org/eclipse/hono/adapter/http/impl/VertxBasedHttpProtocolAdapter.java
@@ -11,7 +11,7 @@
  * SPDX-License-Identifier: EPL-2.0
  *******************************************************************************/
 
-package org.eclipse.hono.adapter.http.vertx;
+package org.eclipse.hono.adapter.http.impl;
 
 import java.net.HttpURLConnection;
 import java.util.Objects;
