+++ adapters/kura/src/main/java/org/eclipse/hono/adapter/kura/impl/Application.java
@@ -10,7 +10,8 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  *******************************************************************************/
-package org.eclipse.hono.adapter.http.vertx;
+
+package org.eclipse.hono.adapter.kura.impl;
 
 import org.eclipse.hono.service.AbstractApplication;
 import org.springframework.boot.SpringApplication;
@@ -19,17 +20,19 @@ import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 
 /**
- * The Hono HTTP adapter main application class.
+ * The Kura adapter main application class.
  */
-@ComponentScan(basePackages = { "org.eclipse.hono.adapter.http", "org.eclipse.hono.service.metric" })
+@ComponentScan("org.eclipse.hono.adapter.kura.impl")
+@ComponentScan("org.eclipse.hono.service.credentials")
+@ComponentScan("org.eclipse.hono.service.metric")
 @Configuration
 @EnableAutoConfiguration
 public class Application extends AbstractApplication {
 
     /**
-     * Starts the HTTP Adapter application.
+     * Starts the Kura Adapter application.
      * 
-     * @param args Command line args passed to the application.
+     * @param args Command line arguments passed to the application.
      */
     public static void main(final String[] args) {
         SpringApplication.run(Application.class, args);
