+++ adapters/mqtt-vertx/src/main/java/org/eclipse/hono/adapter/mqtt/impl/Application.java
@@ -22,7 +22,8 @@ import org.springframework.context.annotation.Configuration;
 /**
  * The Hono MQTT adapter main application class.
  */
-@ComponentScan(basePackages = { "org.eclipse.hono.adapter.mqtt", "org.eclipse.hono.service.metric" })
+@ComponentScan("org.eclipse.hono.adapter.mqtt")
+@ComponentScan("org.eclipse.hono.service.metric")
 @Configuration
 @EnableAutoConfiguration
 public class Application extends AbstractApplication {
